requirements.txt
python-dotenv==1.0.1
scipy>=1.12.0
fastapi==0.110.0
uvicorn==0.27.1
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
joblib==1.3.2
pydantic==2.6.4
kafka-python==2.0.2
requests==2.32.3
pytest==9.0.3
